+++ src/predict/__init__.py
@@ -1,0 +1,33 @@
+# Copyright (c) 2026 Pl4yer-ONE
+# This file is part of FragAudit.
+# Licensed under GPLv3.
+
+"""
+Prediction Module
+Round win probability and player impact forecasting.
+"""
+
+from .win_model import (
+    RoundFeatures,
+    RoundPrediction,
+    WinPredictor,
+    predict_round_win,
+)
+
+from .player_model import (
+    PlayerFeatures,
+    PlayerPrediction,
+    ImpactPredictor,
+    predict_player_impact,
+)
+
+__all__ = [
+    'RoundFeatures',
+    'RoundPrediction',
+    'WinPredictor',
+    'predict_round_win',
+    'PlayerFeatures', 
+    'PlayerPrediction',
+    'ImpactPredictor',
+    'predict_player_impact',
+]